refactor: Study api에서 Participant api 분리
#89
