+++ src/main/java/doore/study/api/StudyController.java
@@ -80,29 +80,4 @@ public class StudyController {
         studyCommandService.terminateStudy(studyId);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
-
-    @PostMapping("/studies/{studyId}/members/{memberId}")
-    public ResponseEntity<Void> saveParticipant(@PathVariable Long studyId, @PathVariable Long memberId) {
-        studyCommandService.saveParticipant(studyId, memberId);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-    @DeleteMapping("/studies/{studyId}/members/{memberId}")
-    public ResponseEntity<Void> deleteParticipant(@PathVariable Long studyId, @PathVariable Long memberId) {
-        studyCommandService.deleteParticipant(studyId, memberId);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
-
-    @DeleteMapping("/studies/{studyId}/members")
-    public ResponseEntity<Void> withdrawParticipant(@PathVariable Long studyId, HttpServletRequest request) {
-        //Todo: 이후 권한 로직으로 수정
-        studyCommandService.withdrawParticipant(studyId, request);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
-    }
-
-    @GetMapping("/studies/{studyId}/members")
-    public ResponseEntity<List<Participant>> getParticipant(@PathVariable Long studyId) {
-        List<Participant> participants = studyQueryService.findAllParticipants(studyId);
-        return ResponseEntity.ok(participants);
-    }
 }
