+++ src/main/java/doore/study/application/StudyCommandService.java
@@ -36,8 +36,6 @@ public class StudyCommandService {
     private final StudyRepository studyRepository;
     private final TeamRepository teamRepository;
     private final CurriculumItemRepository curriculumItemRepository;
-    private final ParticipantRepository participantRepository;
-    private final MemberRepository memberRepository;
 
     public void createStudy(final StudyCreateRequest request, final Long teamId) {
         validateExistTeam(teamId);
@@ -104,42 +102,11 @@ public class StudyCommandService {
         }
     }
 
-    public void saveParticipant(Long studyId, Long memberId) {
-        validateExistStudy(studyId);
-        Member member = validateExistMember(memberId);
-        Participant participant = Participant.builder()
-                .studyId(studyId)
-                .member(member)
-                .build();
-        participantRepository.save(participant);
-    }
-
-    public void deleteParticipant(Long studyId, Long memberId) {
-        validateExistStudy(studyId);
-        Member member = validateExistMember(memberId);
-        participantRepository.deleteByStudyIdAndMember(studyId, member);
-    }
-
-    public void withdrawParticipant(Long studyId, HttpServletRequest request) {
-        validateExistStudy(studyId);
-        String memberId = request.getHeader("Authorization");
-        if (memberId == null) {
-            throw new MemberException(UNAUTHORIZED);
-        }
-        Member member = memberRepository.findById(Long.parseLong(memberId))
-                .orElseThrow(() -> new MemberException(NOT_FOUND_MEMBER));
-        participantRepository.deleteByStudyIdAndMember(studyId, member);
-    }
-
     private void validateExistTeam(Long teamId) {
         teamRepository.findById(teamId).orElseThrow(() -> new TeamException(TeamExceptionType.NOT_FOUND_TEAM));
     }
 
     private Study validateExistStudy(Long studyId) {
         return studyRepository.findById(studyId).orElseThrow(() -> new StudyException(NOT_FOUND_STUDY));
-    }
-
-    private Member validateExistMember(Long memberId) {
-        return memberRepository.findById(memberId).orElseThrow(() -> new MemberException(NOT_FOUND_MEMBER));
     }
 }
