+++ src/test/java/doore/restdocs/RestDocsTest.java
@@ -8,6 +8,8 @@ import doore.helper.ApiTestHelper;
 import doore.login.application.LoginService;
 import doore.member.application.MemberCommandService;
 import doore.study.application.CurriculumItemCommandService;
+import doore.study.application.ParticipantCommandService;
+import doore.study.application.ParticipantQueryService;
 import doore.study.application.StudyCommandService;
 import doore.study.application.StudyQueryService;
 import doore.team.application.TeamCommandService;
@@ -43,6 +45,12 @@ public abstract class RestDocsTest extends ApiTestHelper {
 
     @MockBean
     protected StudyQueryService studyQueryService;
+
+    @MockBean
+    protected ParticipantCommandService participantCommandService;
+
+    @MockBean
+    protected ParticipantQueryService participantQueryService;
 
     @MockBean
     protected CurriculumItemCommandService curriculumItemCommandService;
